# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Owner-scoped tasks, identity, wallet, sandbox payments and notifications.",
        "title": "FinTask API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v2/auth/login": {
            "post": {
                "description": "Создает серверную cookie-сессию и отдельный CSRF-токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "Войти по email и паролю",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вход выполнен",
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/logout": {
            "post": {
                "security": [
                    {
                        "CSRFToken": [],
                        "SessionCookie": []
                    }
                ],
                "description": "Отзывает серверную сессию и очищает session/CSRF cookies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "Завершить текущую сессию",
                "responses": {
                    "204": {
                        "description": "Сессия завершена"
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректный CSRF-токен",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/password-reset/confirm": {
            "post": {
                "description": "Потребляет одноразовый токен, меняет пароль и отзывает старые сессии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "Установить новый пароль",
                "parameters": [
                    {
                        "description": "Токен и новый пароль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.ConfirmPasswordResetRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Пароль изменен"
                    },
                    "400": {
                        "description": "Токен неверен, просрочен или использован",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/password-reset/request": {
            "post": {
                "description": "Создает одноразовый токен, если активный аккаунт существует",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "Запросить сброс пароля",
                "parameters": [
                    {
                        "description": "Email аккаунта",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.RequestPasswordResetRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Запрос принят",
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.RequestPasswordResetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/register": {
            "post": {
                "description": "Создает аккаунт и отправляет запрос на подтверждение email через outbox",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "Зарегистрировать аккаунт",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.RegisterAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Аккаунт ожидает подтверждения email",
                        "schema": {
                            "$ref": "#/definitions/internal_features_identity_transport_http.RegisterAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
...
        "internal_features_identity_transport_http.ConfirmPasswordResetRequest": {
            "type": "object",
            "required": [
                "new_password",
                "token"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 12
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_features_identity_transport_http.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 12,
                    "example": "strong-password"
                }
            }
        },
        "internal_features_identity_transport_http.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "user_id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "internal_features_identity_transport_http.RegisterAccountRequest": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "example": "user@example.com"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Ivan Ivanov"
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 12,
                    "example": "strong-password"
                }
            }
        },
        "internal_features_identity_transport_http.RegisterAccountResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "status": {
                    "type": "string",
                    "example": "pending_verification"
                },
                "user_id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "internal_features_identity_transport_http.RequestPasswordResetRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "example": "user@example.com"
                }
            }
        },
        "internal_features_identity_transport_http.RequestPasswordResetResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access JWT в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRFToken": {
            "description": "CSRF-токен для изменяющих запросов с cookie-сессией",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        },
        "SessionCookie": {
            "description": "Защищенная cookie-сессия браузера",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}